Renamed rubygems dependency to external dependency
